examples/uicopilot/pyproject.toml
[project]
name = "uicopilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.6.0",
    "anthropic>=0.50.0",
    "google-genai>=1.11.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]

[project.optional-dependencies]
training = []

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "ruff>=0.11.6",
    "setuptools>=79.0.1",
    "types-requests>=2.32.0.20250328",
]

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
# disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 150
exclude = [".venv", "venv", "__pycache__", ".git"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]
ignore = ["E501", "W293"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.mccabe]
max-complexity = 5
